refactor(utils/RouterInfo): /utils/api -> /utils/loader로 변경
