+++ src/utils/loader/FriendsListLoader.jsx
